create a work-around for the wrong-url-connection problem
When an HttpURLConnection is created instead of an expected HttpsURLConnection, forcibly replace the handler with a handler that will create the right kind of connection.

For now I am just using the oracle JRE's https URLStreamHandler implementation, and not peoplesoft's. I don't know enough about it to know if this is problematic.