debug-print the location of the wsdl document
Recently, we hit a problem where the PS servers were using wsapi staging for the WSDL,
even though the soap requests were correctly being sent to wsapi production.

The fault was with the wsapi-postalsoft-client jar; it was an older one that had the wsdlLocation attribute set to https://wsapistaging.ccci.org..., and though I had at some point updated the wsapi config to not do this anymore, I must not have ran 'mvn install' or something.
I gave an updated jar to the PSHR admin to fix this.

This new debug printing proves that the new jar is working correctly,
and not sending wsdl requests to wsapi staging.
(It's now getting the wsdl from the jar.)